+++ src/datasources/CustomTwigTemplate.php
@@ -18,7 +18,7 @@ class CustomTwigTemplate extends DataSource
     /**
      * @inheritdoc
      */
-    public function getName(): string
+    public static function displayName(): string
     {
         return Craft::t('sprout-reports', 'Twig Template');
     }